Move stories examples to examples/ directory
Also, the dependencies to run the stories have been moved to
stories/package.json and can be installed with npm run build-storybook.
