+++ stories/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "stories",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "autosuggest-highlight": "^3.1.1",
+    "react": "^16.4.1",
+    "react-autosuggest": "^9.3.4"
+  }
+}
